// packages/sdk/swagger/models/NEXUZR004OffersSpecialService.ts
export interface NEXUZR004OffersSpecialService {
    /**
     * Logical flight ID.
     * @type {number}
     * @memberof NEXUZR004OffersSpecialService
     */
    lfid: number;
    /**
     * Amount of the special service.
     * @type {number}
     * @memberof NEXUZR004OffersSpecialService
     */
    amount: number;
    /**
     * Currency of the amount.
     * @type {string}
     * @memberof NEXUZR004OffersSpecialService
     */
    currency: string;
    /**
     * Time limit (hrs) to purchase the special service before flight departure.
     * @type {number}
     * @memberof NEXUZR004OffersSpecialService
     */
    cutOffHours: number;
    /**
     * Description of the special service.
     * @type {string}
     * @memberof NEXUZR004OffersSpecialService
     */
    description: string;
    /**
     * Maximum available quantity of the special service.
     * @type {number}
     * @memberof NEXUZR004OffersSpecialService
     */
    maxCountServiceLevel: number;
    /**
     * Current available quantity of the special service.
     * @type {number}
     * @memberof NEXUZR004OffersSpecialService
     */
    qtyAvailable: number;
    /**
     * SSR code of the special service.
     * @type {string}
     * @memberof NEXUZR004OffersSpecialService
     */
    ssrCode: string;
    /**
     * SSR ID of the special service.
     * @type {number}
     * @memberof NEXUZR004OffersSpecialService
     */
    ssrId: number;
    /**
     * Number of days until sales start.
     * @type {number}
     * @memberof NEXUZR004OffersSpecialService
     */
    startSalesDays: number;
    /**
     * Physical flight ID.
     * @type {number}
     * @memberof NEXUZR004OffersSpecialService
     */
    pfid?: number;
}